ci: use personal access token for github
semantic-release@15 doesn't work with the token provided by github
actions by default (see semantic-release/semantic-release#1194).
A fix is provided with semantic-release@16 which is in beta at the
moment and breaks npm releases here.
Work around this by using a personal access token provided
explicitly to the actions in form of a secret.